:recycle: refactor: refactoring server.js
